+++ ext/stub/bench.zep.h
@@ -38,6 +38,10 @@ PHP_METHOD(Stub_Bench, countUnoptimizedLoop);
 PHP_METHOD(Stub_Bench, implodeOptimizedLoop);
 PHP_METHOD(Stub_Bench, implodeUnoptimizedLoop);
 PHP_METHOD(Stub_Bench, acosLoop);
+PHP_METHOD(Stub_Bench, conditionalLocalLoop);
+PHP_METHOD(Stub_Bench, conditionalLocalStep);
+PHP_METHOD(Stub_Bench, definiteLocalLoop);
+PHP_METHOD(Stub_Bench, definiteLocalStep);
 void zep_Stub_Bench_zephir_gen_step_generatorRange(int ht, zend_execute_data *execute_data, zval *return_value, zval *this_ptr, int return_value_used, zval *zephir_gen_ext );
 void zep_Stub_Bench_zephir_gen_step_generatorOverArray(int ht, zend_execute_data *execute_data, zval *return_value, zval *this_ptr, int return_value_used, zval *zephir_gen_ext );
 
@@ -187,6 +191,22 @@ ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_stub_bench_acosloop, 0, 1, IS_DO
 	ZEND_ARG_TYPE_INFO(0, n, IS_LONG, 0)
 ZEND_END_ARG_INFO()
 
+ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_stub_bench_conditionallocalloop, 0, 1, IS_LONG, 0)
+	ZEND_ARG_TYPE_INFO(0, iterations, IS_LONG, 0)
+ZEND_END_ARG_INFO()
+
+ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_stub_bench_conditionallocalstep, 0, 1, IS_LONG, 0)
+	ZEND_ARG_TYPE_INFO(0, flag, _IS_BOOL, 0)
+ZEND_END_ARG_INFO()
+
+ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_stub_bench_definitelocalloop, 0, 1, IS_LONG, 0)
+	ZEND_ARG_TYPE_INFO(0, iterations, IS_LONG, 0)
+ZEND_END_ARG_INFO()
+
+ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_stub_bench_definitelocalstep, 0, 1, IS_LONG, 0)
+	ZEND_ARG_TYPE_INFO(0, flag, _IS_BOOL, 0)
+ZEND_END_ARG_INFO()
+
 ZEND_BEGIN_ARG_INFO_EX(arginfo_stub_bench_zephir_gen_step_generatorrange, 0, 0, 1)
 	ZEND_ARG_INFO(0, zephir_gen)
 ZEND_END_ARG_INFO()
@@ -231,5 +251,9 @@ PHP_ME(Stub_Bench, __construct, arginfo_stub_bench___construct, ZEND_ACC_PUBLIC|
 	PHP_ME(Stub_Bench, implodeOptimizedLoop, arginfo_stub_bench_implodeoptimizedloop, ZEND_ACC_PUBLIC)
 	PHP_ME(Stub_Bench, implodeUnoptimizedLoop, arginfo_stub_bench_implodeunoptimizedloop, ZEND_ACC_PUBLIC)
 	PHP_ME(Stub_Bench, acosLoop, arginfo_stub_bench_acosloop, ZEND_ACC_PUBLIC)
+	PHP_ME(Stub_Bench, conditionalLocalLoop, arginfo_stub_bench_conditionallocalloop, ZEND_ACC_PUBLIC)
+	PHP_ME(Stub_Bench, conditionalLocalStep, arginfo_stub_bench_conditionallocalstep, ZEND_ACC_PUBLIC)
+	PHP_ME(Stub_Bench, definiteLocalLoop, arginfo_stub_bench_definitelocalloop, ZEND_ACC_PUBLIC)
+	PHP_ME(Stub_Bench, definiteLocalStep, arginfo_stub_bench_definitelocalstep, ZEND_ACC_PUBLIC)
 	PHP_FE_END
 };